Fallback local language to en if not supported
SOLVES https://github.com/oam-dev/velaux/issues/488

Currently only en and zh are the only supported languages.
If for example in reason of other browser language settings, the locales doesn't exist and the ui is not visible (after login).
As workaround it is possible to change the browser language (bad way).

The solution is to add a fallback language.

Signed-off-by: Manuel Ruck <git@manuelruck.de>